Prepared scenario for multiple configuration background images
Added helper function that can find attachment images with a specific name so that when adding new images, only those are remaning.
While it is still unsupported, the program already detects all of the added background images correctly, now what remains to do is to modify the preview and see what images one has added, create the actual slideshow, and an option to modify the time for each image.