Added slider and toggle for background image compression
Now there are two values in the database, one for activating and another one that sets the compression factor for the background images.
Note that these values yet have to be implemented in the image library.

Also, the extended configuration only activated for the custom images section has been modified to only show up when a valid background image mode has been selected in order to declutter the UI when those options are unavailable.