Added uncompressed and compressed images to all metadata files for games
Now the game accepts either uncompressed and compressed images for all of the types of image types that sharp can accept. Granted, there is still work to be done, mainly with error reporting and making some changes onto other places like when cropping images.

To avoid users having to deal with the cropping automatically when they add a new image, said window has been disabled for when adding a new image, since there's already an option to edit them manually.